.github/scripts/pyproject.toml
[tool.poetry]
name = "gh_scripts"
version = "0.0.1"
description = ""
authors = ["Palo Alto Networks"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
requests = "^2.31"
